requirements.txt
flask
pdfminer.six
python-docx
scikit-learn
nltk
joblib
regex
pillow
pytesseract
